Tests for character movement, fixed acceleration calculations and changed physics to use doubles

+++ game/src/main/java/fi/jakojaannos/roguelite/game/data/components/CharacterInput.java
@@ -1,8 +1,8 @@
 package fi.jakojaannos.roguelite.game.data.components;
 
 import fi.jakojaannos.roguelite.engine.ecs.Component;
-import org.joml.Vector2f;
+import org.joml.Vector2d;
 
 public class CharacterInput implements Component {
-    public Vector2f move = new Vector2f(0.0f, 0.0f);
+    public Vector2d move = new Vector2d(0.0f, 0.0f);
 }
