+++ game/src/main/java/fi/jakojaannos/roguelite/game/data/components/Position.java
@@ -3,10 +3,10 @@ package fi.jakojaannos.roguelite.game.data.components;
 import fi.jakojaannos.roguelite.engine.ecs.Component;
 
 public class Position implements Component {
-    public float x;
-    public float y;
+    public double x;
+    public double y;
 
-    public Position(float x, float y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
